Inverted Sections
Adds a new sigil, the hat (^), which can be used to denote
an Inverted Section. From mustache(5):

An inverted section begins with a caret (hat) and ends with a
slash. That is `{{^person}}` begins a "person" inverted section while
`{{/person}}` ends it.

While sections can be used to render text one or more times based on the
value of the key, inverted sections may render text once based
on the inverse value of the key. That is, they will be rendered
if the key doesn't exist, is false, or is an empty list.

Template:

    {{#repo}}
      <b>{{name}}</b>
    {{/repo}}
    {{^repo}}
      No repos :(
    {{/repo}}

Hash:

    {
      "repo": []
    }

Output:

    No repos :(